change adapter little bit
